tdr-extractor/pyproject.toml
[tool.poetry]
name = "tdr-extractor"
version = "0.1.0"
description = "Microservicio para extracción de campos desde PDFs de TDR usando OCR y LLM"
authors = ["Asiter Team"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
pypdf = "^5.1.0"
pytesseract = "^0.3.13"
pillow = "^11.0.0"
python-multipart = "^0.0.17"
pdf2image = "^1.17.0"
httpx = "^0.28.0"
pydantic = "^2.10.0"
pydantic-settings = "^2.6.0"
python-dotenv = "^1.0.1"

# Para modelo local con llama-cpp (opcional, descomentar si se usa)
# llama-cpp-python = "^0.3.0"

# Para usar DeepSeek-OCR via API
openai = "^1.55.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
pytest-asyncio = "^0.24.0"
requests = "^2.32.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"

